milliondolarproject/lgs-platform/backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
alembic==1.11.1
pydantic==2.5.2
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.0
httpx==0.26.0
pdfplumber==0.11.8
PyPDF2==3.0.1
anthropic==0.72.1
pillow>=10.0.0
pypdfium2>=4.0.0

